Bug found for this new construction.
It's not enough to remember a single boolean in this case, because when you
fail a progress check, you don't know what should be the boolean for the
previous star you were in (in case of nested stars).
So we would need to keep track of one boolean for each nested star, increasing
the worst-case size of the state space to r².

This documents several counter-examples (found by fuzzer, and minimal by hand).